requirements.txt
attrs==23.1.0
barcode==1.0.4
blinker==1.7.0
cachelib==0.12.0
certifi==2023.7.22
cffi==1.15.1
click==8.1.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
Flask==3.0.2
Flask-Session==0.8.0
h11==0.14.0
idna==3.4
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
Levenshtein==0.25.0
MarkupSafe==2.1.5
msgspec==0.18.6
openpyxl==3.1.2
outcome==1.2.0
Pillow==10.0.0
pycparser==2.21
pyperclip==1.8.2
PySocks==1.7.1
python-barcode==0.15.1
python-Levenshtein==0.25.0
pyzbar==0.1.9
rapidfuzz==3.7.0
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.10.0
urllib3==2.0.4
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.18.1
